requirements.txt
aiogram==3.0.0b6
python-dotenv==1.0.0
pillow==10.0.0
